move WireGuardDatagramTransport into dedicated module
this allows us to exclude it from individual coverage, which makes no sense.
Also improve type checking to make sure that it's a full replacement.
