+++ setup.cfg
@@ -69,6 +69,7 @@ exclude =
     mitmproxy/net/http/message.py
     mitmproxy/net/http/multipart.py
     mitmproxy/net/tls.py
+    mitmproxy/net/udp_wireguard.py
     mitmproxy/options.py
     mitmproxy/proxy/config.py
     mitmproxy/proxy/server.py
