+++ test/mitmproxy/net/test_udp_wireguard.py
@@ -1,0 +1,1 @@
+# testing this in isolation makes no sense. See proxy/test_mode_servers.py.
